refactor(generate): separate jsdoc parsing into its own package
Closes #108
